// src/types.ts
export type ViewXY = {
  viewX: number;
  viewY: number;
};

export type ViewSize = {
  viewWidth: number;
  viewHeight: number;
};

export type WorldXY = {
  worldX: number;
  worldY: number;
};

export type OffsetXY = {
  offsetX: number;
  offsetY: number;
};

export type RGB = [number, number, number];